backend/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.13.2
altair==5.4.0
anthropic>=0.52.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
assemblyai==0.33.0
asttokens==2.4.1
attrs==24.1.0
audioread==3.0.1
av==12.0.0
backoff==2.2.1
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
croniter==1.4.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepgram-sdk==4.8.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
executing==2.0.1
fal_client==0.4.1
fastapi==0.112.0
fastapi-cli==0.0.5
fastapi-utilities==0.2.0
filelock==3.15.4
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
graphviz==0.20.3
groq==0.9.0
grpcio==1.65.4
grpcio-status==1.62.3
grpcio-tools==1.62.3
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jiwer==3.0.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-pinecone==0.2.12
pinecone==7.3.0
# Self-host vector DB (pgvector). Only used when VECTOR_DB=pgvector.
psycopg[binary,pool]>=3.2
pgvector>=0.3
# Self-host object storage (MinIO, S3-compatible). Only used when STORAGE_BACKEND=minio.
boto3>=1.34
# Self-host document store (MongoDB). Only used when DB_BACKEND=mongo.
pymongo>=4.6
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-sdk==0.2.9
langsmith==0.4.37
lazy_loader==0.4
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==10.5.0
mplcursors==0.5.3
mpld3==0.5.10
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.5.2
neo4j==5.23.1
networkx==3.3
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.19.0
openai==1.104.2
scipy==1.14.0
optuna==3.6.1
opuslib==3.0.1
orjson==3.10.6
prometheus-client==0.21.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
polling2==0.5.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.5.2
primePy==1.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.10.1
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pynndescent==0.5.13
PyOgg @ git+https://github.com/TeamPyOgg/PyOgg@6871a4f234e8a3a346c4874a12509bfa02c4c63a
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
python-ulid==3.0.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.11.0
rapidfuzz==3.9.7
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests~=2.32.5
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
semver==3.0.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SoundCard==0.4.3
soundfile==0.12.1
soxr==0.4.0
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.1
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.3
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
webrtcvad==2.0.10
websockets==12.0
yarl==1.9.4
stripe==11.3.0
typesense==0.21.0
pycountry==24.6.1
google-cloud-translate==3.20.2
langdetect==1.0.9
lc3py==1.1.3
twilio==9.5.0
